Sprint 2 (#589 D1): split RecruitmentCandidateRepository into Reader + Writer
Extends the A6 read/write split to the largest repository left outside #563.
Static-repo pattern (StaticRepositoryTrait), mirroring the CustomFieldRepository
reference:

- RecruitmentCandidateReader — the 14 reads (get_by_id/ids, the *_hash lookups,
  get_by_user_id, the paginated/filtered listings + their counts,
  get_ids_by_email_hash) and get_table_name.
- RecruitmentCandidateWriter — create, update, set_user_id, delete.
- RecruitmentCandidateRepository stays the public façade delegating every method,
  keeping its constants and the @phpstan-type (CandidateRow). Reader/Writer share
  the 'ffc_recruitment_candidate' cache group so writes still invalidate read
  caches; get_table_name is duplicated into both. Tech-debt note (#563 B3) added.

The repo test gains @covers for the Reader/Writer so pcov attributes the
delegated coverage (verified: Reader/Writer land coverage), protecting the floor.

Behavior-preserving — Recruitment suite green (566 tests); PHPStan L8 + WPCS clean.
